pyproject.toml
[project]
name = "forkchat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "databases>=0.9.0",
    "fastapi>=0.115.12",
    "motor>=3.7.1",
    "passlib>=1.7.4",
    "psycopg2>=2.9.10",
    "pydantic[email]>=2.11.4",
    "pytest>=8.3.5",
    "python-jose>=3.4.0",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.2",
]
